+++ packages/terra-embedded-content-consumer/tests/jest/EmbeddedContentConsumer.test.jsx
@@ -81,3 +81,31 @@ it('should validate the inputs', () => {
   expect(wrapper.instance().props.options.resizeConfig).toBe(resizeConfig);
   expect(wrapper.instance().props.eventHandlers).toBe(customEvents);
 });
+
+it('sets appropriate config option when resizeConfig.scrolling is true', () => {
+  const embeddedContentConsumer = (
+    <div>
+      <EmbeddedContentConsumer
+        src="https://www.google.com/"
+        options={{ resizeConfig: { scrolling: true } }}
+      />
+    </div>
+  );
+
+  const wrapper = shallow(embeddedContentConsumer);
+  expect(wrapper).toMatchSnapshot();
+});
+
+it('sets appropriate config option when resizeConfig.scrolling is false', () => {
+  const embeddedContentConsumer = (
+    <div>
+      <EmbeddedContentConsumer
+        src="https://www.google.com/"
+        options={{ resizeConfig: { scrolling: false } }}
+      />
+    </div>
+  );
+
+  const wrapper = shallow(embeddedContentConsumer);
+  expect(wrapper).toMatchSnapshot();
+});