+++ packages/terra-framework-docs/src/terra-dev-site/doc/embedded-content-consumer/example/BasicConsumerWithScrolling.jsx
@@ -1,18 +1,17 @@
 import React from 'react';
 import EmbeddedContentConsumer from 'terra-embedded-content-consumer';
 import { Consumer } from 'xfc';
-import '../providers/EmbeddedContentConsumerCommon.module.scss';
 
 Consumer.init();
 
-const BasicConsumer = () => (
+const BasicConsumerWithScrolling = () => (
   <EmbeddedContentConsumer
     src="/terra-framework/#/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/basic-provider"
     options={{
-      iframeAttrs: { title: 'Basic content example' },
-      resizeConfig: { scrolling: true },
+      iframeAttrs: { title: 'Basic content example of scrolling is enabled' },
+      resizeConfig: { scrolling: true, fixedWidth: '100%', fixedHeight: '100px' },
     }}
   />
 );
 
-export default BasicConsumer;
+export default BasicConsumerWithScrolling;