+++ packages/terra-embedded-content-consumer/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added an example and updated documentation on how to enable embedded content to be scrolled within the iframe.
+
 ## 3.38.0 - (September 26, 2023)
 
 * Updated